switch: don't accidentally sign-extend trigger flags
God, I sure hate these vague cfile functions. (I'm probably going to
clean them up all at once one day, if I live that long.)

This affects only d2x-xl levels, since the flag value 128 (TF_AUTOPLAY)
is a d2x-xl extension.

+++ main/switch.c
@@ -649,7 +649,7 @@ extern void trigger_read(trigger *t, CFILE *fp, bool obj_trigger)
 	if (obj_trigger) {
 		t->flags = (uint16_t)cfile_read_short(fp);
 	} else {
-		t->flags = cfile_read_byte(fp);
+		t->flags = (uint8_t)cfile_read_byte(fp);
 	}
 	t->num_links = cfile_read_byte(fp);
 	cfile_read_byte(fp);
